Update README.md
A store sells coffee for $8.50 per pound. If a customer buys more than 3 pounds of coffee, they get a discount of 12% on their entire purchase.

Write a program that inputs how many pounds of coffee that a user wishes to buy and then calculates and outputs the total cost of the coffee.